Make validations open and extensible
New validations may be added from any part of the code.  Validations are run
in no particular order, all must pass without throwing exceptions.

Movement defines a validation for blocking moves by checking if more than one
object is present on a field that are "movement blockers".  Wall and Being are
are denoted as movement blockers because they derive from `movement/blocker`.
This can be defined anywhere in code.

The next interesting question is how to model objects that are sometimes
blockers and sometimes not (like closed vs. open doors), or beings that are no
movement blockers, like ghosts.