Complete client-server step
The eariler "server" commit was sadly missing some critical pieces.

Here we abuse the REPL renderer on the ClojureScript side.  For that to work,
most of the shared code needed to be put in *.cljc files.  Some conditional
compliation was also necessary at this stage.  Hopefully we'll find a cleaner
way later.